重构了数值模块，TimeScale由数值组件来维护，更新TimeScale时会调用NumericChangeCallback

+++ Unity/Assets/Scripts/Loader/Timeline/TimelinePlayer.cs
@@ -23,7 +23,7 @@ namespace Timeline
     public struct UpdateHertzCallback
     {
         public long instanceId;
-        public float Hertz;
+        public int Hertz;
     }
     #endregion
